UPDATE: organizes documents
